Main: move ComPtr to D3D11 RenderSystem which is the only consumer
